Document FixedPoint acos bug
More generally, we should be wary of FixedPoint functions that are not continuous or that are defined on a bounded domain, as they are likely to run into similar bugs in edge cases.

Perhaps it makes sense to start a "known bugs" list soon?

The implementation of sqrt was also changed to prevent a possible similar bug, and also a known bug to do with edge cases having to do with offsets. See the related changes to mod2Pi for why this is a problem. Basically, a number could be equal to FixedPoint.Zero but have negative BigDecimal representation, which would break BigDecimalMath.