Cache sine and cosine
This shouldn't be a surprise, but the BigDecimalMath functions can be
expensive. We were repeatedly recomputing the sine and cosine of
rotation angles, but it makes much more sense to store them in the Move
class.

Note that we don't use the rotate(angle) functions anymore, so they
could be removed.