Trait for all puzzles in a particular geometry
This is rather minimal, and might need some reworking.

The methods flatCuts and emptySortedArcs are interesting in that they
are very geometry dependent, but also pretty specific to Puzzle. For
this reason, I haven't put them in Geometry.

Is there a sensible way to split the flatCuts method? The corresponding
method which exists in Puzzle was a bit more detailed, but also quite
implementation-dependent.

Should the lazy vals in this class be vals or even defs? I can imagine
running into problems when extending from Puzzle.

I haven't checked if all of the comments still make sense. I should
also verify that no critical comments were removed.