Increase FixedPoint computing scale
I noticed that it wasn't high enough for some spherical puzzle
calculations, which was resulting in crashes. A better solution might
have been to just set CompareScale lower, but this is safer.

On the whole, these really should be viewed as tunable parameters.
Perhaps there should be a way to set them at runtime?